Fix for Use Clang 16 toolchain
ANDROID: kbuild: Wrap zero trivial-auto-var-init enablement flag with cc-option

Starting with LLVM 16, this flag is deprecated, which means its presence
breaks the build:

  Cannot use CONFIG_CC_STACKPROTECTOR_STRONG: -fstack-protector-strong not supported by compiler
  make[2]: *** [../Makefile:1289: prepare-compiler-check] Error 1
  clang-16: warning: the flag '-enable-trivial-auto-var-init-zero-knowing-it-will-be-removed-from-clang' has been deprecated and will be ignored [-Wunused-command-line-argument]
  clang-16: warning: the flag '-enable-trivial-auto-var-init-zero-knowing-it-will-be-removed-from-clang' has been deprecated and will be ignored [-Wunused-command-line-argument]

Wrap this option in cc-option so that it only gets added to the list of
flags when it is supported. This is resolved properly upstream already
with commit 607e57c6c62c ("hardening: Remove Clang's enable flag for
-ftrivial-auto-var-init=zero") but that cannot be easily backported due
to lack of cc-option support in Kconfig, hence this downstream only
change.

Change-Id: Ie1f7bcdec23f01bf19c68b4dae460923aaa150c9
Fixes: a7eda6eddca5 ("BACKPORT: security: allow using Clang's zero initialization for stack variables")
Link: https://github.com/llvm/llvm-project/commit/aef03c9b3bed5cef5a1940774b80128aefcb4095
Signed-off-by: Nathan Chancellor <nathan@kernel.org>