ksu: fs: add various ioctls to ext4 and f2fs This commit adds the following ioctls to ext4 and f2fs: - FS_IOC_GET_ENCRYPTION_POLICY_EX - FS_IOC_ADD_ENCRYPTION_KEY - FS_IOC_REMOVE_ENCRYPTION_KEY - FS_IOC_REMOVE_ENCRYPTION_KEY_ALL_USERS - FS_IOC_GET_ENCRYPTION_KEY_STATUS
In particular, FS_IOC_ADD_ENCRYPTION_KEY and FS_IOC_REMOVE_ENCRYPTION_KEY
will soon be mandatory on Android:
https://android-review.googlesource.com/c/platform/system/vold/+/2834717

'Therefore, this commit removes the code that added and removed keys
to/from the session keyring.  Now the ioctls are used unconditionally.'

Unfortunately, cherry-picking the commits that actually add support for
these ioctls was becoming a much bigger job than just manually copying
the code. This is due to the fact that this kernel is based on an old
CAF tag, and recent TAGs have rebased the code, while we moved forward
following Linux 4.14 LTS branch.

Test performed:
- boot/reboot with encrypted and not-encrypted ext4/f2fs
- upgrade from previous kernels (without these ioctls) with encrypted and
  not-encrypted ext4/f2fs

Note: if encryption is enabled, this only supports ICE/PRIVATE encryption
mode for file contents.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>